Extract metrics into separate file
